Refactor code
- Placed extra files in `utility` folder.
- Created `db.py` to store the basic database queries/functionalities
- Created `palm.py` to place all functions made for using PaLM API
- Created `scrap.py` to move all scraping functions in their own file
