+++ main.py
@@ -2,194 +2,22 @@ import requests
 import os
 import json
 import asyncio
-from typing import List
-import re
 from datetime import date
-import utility as utils
-import db
-import opportunity as opps
-from opportunity import Opportunity, OpportunityType
+from utility import utils
+import utility.db as db
+import utility.opportunity as opps
 from dotenv import load_dotenv
-
-load_dotenv()  # To obtain keys from the .env file
-
-
-# ----------------- POSTGRES -----------------
-
-TABLE_NAME = os.getenv("DB_TABLE")
-MAX_LIST_LENGTH = 13
+from utility.scrap import (
+    request_github_internship24_data,
+    request_linkedin_data,
+    request_linkedin_internship24_data,
+)
+from utility.palm import gpt_job_analyze
 
 
-def create():
-    """Creates the DB. Only needs to be called once."""
-
-    with db.instantiate_db_connection() as connection:
-        cursor = connection.cursor()
-
-        cursor.execute(
-            f"""CREATE TABLE IF NOT EXISTS {TABLE_NAME}(company TEXT, title TEXT, location TEXT, link TEXT, processed INTEGER DEFAULT 0)"""
-        )
-
-        connection.commit()
-
-
-# ----------------- INTERNSHIP DATA -----------------
-
-
-def request_github_internship24_data() -> List[Opportunity]:
-    """Scrapes Internship Data '24 from Github Repo"""
-
-    url = os.getenv("GH_INTERN24_URL")
-    parse_content = utils.content_parser(url)
-    github_list = []
-    td_elems = parse_content.find_all("tr")
-
-    for cell in td_elems[1:]:
-        if len(github_list) <= MAX_LIST_LENGTH:
-            elements = cell.find_all("td")
-
-            company = elements[0].text
-            title = elements[1].text
-            location = elements[2].text
-            link = elements[3]
-            if "🔒" not in link.text:
-                opportunity = Opportunity(
-                    company,
-                    title,
-                    location,
-                    link.find("a")["href"],
-                    0,
-                    OpportunityType.INTERNSHIP.value,
-                )
-                github_list.append(opportunity)
-
-    return github_list
-
-
-def request_linkedin_internship24_data() -> List[Opportunity]:
-    """Web scrapes Summer '24 Internship Opportunities using LinkedIn"""
-
-    url = os.getenv("LINKEDIN_INTERN_URL")
-
-    parse_content = utils.content_parser(url)
-
-    linkedin_internship_opps = utils.blueprint_opportunity_formatter(
-        parse_content,
-        "base-card relative w-full hover:no-underline focus:no-underline base-card--link base-search-card base-search-card--link job-search-card",
-        "hidden-nested-link",
-        "base-search-card__title",
-        "job-search-card__location",
-        "base-card__full-link",
-        True,
-        MAX_LIST_LENGTH,
-        OpportunityType.INTERNSHIP.value,
-    )
-
-    return linkedin_internship_opps
-
-
-# ----------------- JOB DATA -----------------
-
-
-def request_rapidapi_indeed_data() -> List[Opportunity]:
-    """
-    This API call retrieves a formatted response object
-    and returns a List[Opportunity] as the result
-    """
-
-    url = os.getenv("RAPID_API_URL")
-    rapid_api_key = os.getenv("RAPID_API_KEY")
-
-    headers = {
-        "X-RapidAPI-Key": rapid_api_key,
-        "X-RapidAPI-Host": "indeed12.p.rapidapi.com",
-    }
-
-    rapid_jobs = []
-    response = requests.get(url, headers=headers).json()
-
-    days_needed_command_value = utils.extract_command_value().days_needed[
-        0
-    ]  # Extracts command-line value
-
-    for elem in response["hits"]:
-        time = elem["formatted_relative_time"]
-
-        numeric = re.search(r"\d+", time)
-        formatted_time_integer = int(numeric.group()) if numeric else 0
-
-        if (
-            len(rapid_jobs) < MAX_LIST_LENGTH
-            and int(days_needed_command_value) >= formatted_time_integer
-        ):
-            company = elem["company_name"]
-            title = elem["title"]
-            location = elem["location"]
-            link = f'https://www.indeed.com/viewjob?jk={elem["id"]}&locality=us'
-            processed = 0
-
-            opportunity = Opportunity(
-                company,
-                title,
-                location,
-                link,
-                processed,
-                OpportunityType.FULL_TIME.value,
-            )
-
-            rapid_jobs.append(opportunity)
-
-    return rapid_jobs
-
-
-def request_linkedin_data() -> List[Opportunity]:
-    """Returns a List[Opportunity] which contains web scraped job content"""
-
-    url = os.getenv("LINKEDIN_URL")
-    parse_content = utils.content_parser(url)
-
-    linked_in_jobs = utils.blueprint_opportunity_formatter(
-        parse_content,
-        "base-card relative w-full hover:no-underline focus:no-underline base-card--link base-search-card base-search-card--link job-search-card",
-        "hidden-nested-link",
-        "base-search-card__title",
-        "job-search-card__location",
-        "base-card__full-link",
-        True,
-        MAX_LIST_LENGTH,
-        OpportunityType.FULL_TIME.value,
-    )
-
-    return linked_in_jobs
-
-
-# ----------------- RESET FUNCTION (DEBUGGING PURPOSES) -----------------
-
-
-def reset_processed_status(TABLE_NAME):
-    """Jobs status will be set to _processed = 0 for testing a debugging purposes"""
-
-    with db.instantiate_db_connection() as connection:
-        cursor = connection.cursor()
-
-        cursor.execute(
-            f"SELECT company, title, location FROM {TABLE_NAME} WHERE processed = 1 LIMIT 5"
-        )
-
-        rows = cursor.fetchall()
-
-        for row in rows:
-            company, title, location = row[:3]
-
-            cursor.execute(
-                f"UPDATE {TABLE_NAME} SET processed = 0 WHERE company = %s AND title = %s AND location = %s",
-                (company, title, location),
-            )
-
-        connection.commit()
-
-
-# ----------------- DISCORD BOT -----------------
+# Load and determine if all env variables are set
+load_dotenv()
+utils.verify_set_env_variables()
 
 
 async def execute_opportunities_webhook(webhook_url, job_message, internship_message):
@@ -247,7 +75,10 @@ async def main():
     # Creates table in database
     with_create_table_command = utils.extract_command_value().create
     if with_create_table_command:
-        create()
+        TABLE_NAME = os.getenv("DB_TABLE")
+
+        db.create(TABLE_NAME)
+
         print(f"Sucessfully created {TABLE_NAME}!")
         exit()  # Exit the main function to avoid calling other functions
 
@@ -265,7 +96,7 @@ async def main():
     # Consolidates all job-related opportunities into a comprehensive List[Opportunity], eliminating repetitive calls to the LLM SERVER.
     job_opps = utils.merge_all_opportunity_data(request_linkedin_data())
 
-    filtered_job_opps = utils.gpt_job_analyze(
+    filtered_job_opps = gpt_job_analyze(
         job_opps,
         prompt_object["full_time"],
     )
@@ -277,7 +108,7 @@ async def main():
         request_github_internship24_data(),
     )
 
-    filtered_internship_opps = utils.gpt_job_analyze(
+    filtered_internship_opps = gpt_job_analyze(
         internship_opps,
         prompt_object["internship"],
     )
@@ -289,7 +120,7 @@ async def main():
     # To do so, please comment the function calls above this comment.
     # After, please uncomment the following line of code:
 
-    # reset_processed_status()
+    # db.reset_processed_status()
 
     internship_data_results = opps.list_opportunities(True, "internship", filtered=True)
     job_data_results = opps.list_opportunities(True, "full_time", filtered=True)
@@ -311,5 +142,5 @@ async def main():
     opps.update_opportunities_status(internship_data_results)
 
 
-if __name__ == "__main__":
-    asyncio.run(main())
+# if __name__ == "__main__":
+#     asyncio.run(main())
