+++ utility/opportunity.py
